Refactor: Centralize browser and node target constants
- Moved BROWSER_TARGETS and NODE_TARGET constants to a new shared file for better organization and accessibility.
- Updated imports across various scripts to reference the new constants location.
- Removed redundant definitions from previous locations to streamline the codebase.
