+++ code/core/src/shared/constants/environments-support.ts
@@ -1,0 +1,10 @@
+export const BROWSER_TARGETS = [
+  'chrome131',
+  'edge134',
+  'firefox136',
+  'safari18.3',
+  'ios18.3',
+  'opera117',
+];
+
+export const NODE_TARGET = 'node20';
