+++ scripts/prepare/bundle.ts
@@ -12,8 +12,10 @@ import type { Options } from 'tsup';
 import { build } from 'tsup';
 import type { PackageJson } from 'type-fest';
 
-import { NODE_TARGET } from '../../code/core/scripts/prep';
-import { BROWSER_TARGETS } from '../../code/core/src/builder-manager';
+import {
+  BROWSER_TARGETS,
+  NODE_TARGET,
+} from '../../code/core/src/shared/constants/environments-support';
 import { exec } from '../utils/exec';
 import { dedent, esbuild, nodeInternals } from './tools';
